http://issues.carrot2.org/browse/CARROT-657
AllTokens.type is now short[]. Corrected bitmasks in ITokenTypeAttribute.




git-svn-id: file:///home/dweiss/carrot2/migration/carrot2.svn/trunk@4222 7ff1d41c-760d-0410-a7ff-a3a56f310b35

+++ core/carrot2-algorithm-stc/src/org/carrot2/clustering/stc/STCClusteringAlgorithm.java
@@ -232,7 +232,7 @@ public final class STCClusteringAlgorithm extends ProcessingComponentBase implem
         sb = new GeneralizedSuffixTree.SequenceBuilder();
 
         final int [] tokenIndex = context.allTokens.wordIndex;
-        final int [] tokenType = context.allTokens.type;
+        final short [] tokenType = context.allTokens.type;
         for (int i = 0; i < tokenIndex.length; i++)
         {
             /* Advance until the first real token. */
